requirements.txt
msgpack-python
PyNaCl
